Got map working correctly as well as desktop viewing.
