--- client/components/map.jsx
@@ -1,51 +1,0 @@
-// import React from "react";
-// import { GoogleMap, useJsApiLoader, Marker } from '@react-google-maps/api';
-// import ProfilePage from "../pages/profile-page";
-
-// const containerStyle = {
-//   width: '400px',
-//   height: '400px'
-// };
-
-// const center = {
-//   lat: 33.44656,
-//   lng: -117.61959
-// };
-
-// function MyComponent() {
-//   const { isLoaded } = useJsApiLoader({
-//     id: 'google-map-script',
-//     googleMapsApiKey: process.env.GOOGLE_API_KEY
-//   })
-//   const [map, setMap] = React.useState(null)
-//   const onLoad = React.useCallback(function callback(map) {
-//     const bounds = new window.google.maps.LatLngBounds(center);
-//     map.fitBounds(bounds);
-//     setMap(map)
-//   }, [])
-//   const onUnmount = React.useCallback(function callback(map) {
-//     setMap(null)
-//   }, [])
-
-//   return (
-//     <>
-//       <ProfilePage />
-//       <GoogleMap
-//         mapContainerStyle={containerStyle}
-//         center={center}
-//         zoom={4}
-//         onLoad={onLoad}
-//         onUnmount={onUnmount}
-//       >
-//         { /* Child components, such as markers, info windows, etc. */}
-//         <>
-//         <Marker
-//           position={{ lat: 33.44656, lng: -117.61959 }}
-//         />
-//         </>
-//       </GoogleMap>
-//     </>
-//   )
-// }
-
-// export default React.memo(MyComponent)
