+++ server/public/styles.css
@@ -143,6 +143,6 @@ input[type="file"] {
   background-color: lightgrey;
 }
 
-.map {
-  margin-top: 2rem;
+.profile-content {
+  margin: 1rem;
 }
